+++ src/main/java/kitchenpos/eatinorders/tobe/application/OrderTableService.java
@@ -1,10 +1,10 @@
 package kitchenpos.eatinorders.tobe.application;
 
-import kitchenpos.eatinorders.domain.OrderRepository;
-import kitchenpos.eatinorders.domain.OrderStatus;
 import kitchenpos.eatinorders.tobe.application.dto.request.OrderTableChangeRequest;
 import kitchenpos.eatinorders.tobe.application.dto.request.OrderTableCreateRequest;
 import kitchenpos.eatinorders.tobe.application.dto.response.OrderTableResponse;
+import kitchenpos.eatinorders.tobe.domain.OrderRepository;
+import kitchenpos.eatinorders.tobe.domain.OrderStatus;
 import kitchenpos.eatinorders.tobe.domain.OrderTable;
 import kitchenpos.eatinorders.tobe.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
