+++ src/main/java/kitchenpos/eatinorders/tobe/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.tobe.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
